Rename core_transforms/map/{ParDo Struct -> pardo_struct}
