+++ learning/katas/go/core_transforms/map/lesson-info.yaml
@@ -20,4 +20,4 @@
 content:
 - pardo
 - pardo_onetomany
-- ParDo Struct
+- pardo_struct
