+++ learning/katas/go/core_transforms/map/pardo_struct/cmd/main.go
@@ -13,17 +13,29 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-module mapelements
+package main
 
-go 1.14
-
-require (
-	github.com/apache/beam v2.22.0+incompatible
-	github.com/golang/protobuf v1.4.2 // indirect
-	go.opencensus.io v0.22.3 // indirect
-	golang.org/x/net v0.0.0-20200602114024-627f9648deb9 // indirect
-	golang.org/x/sys v0.0.0-20200610111108-226ff32320da // indirect
-	google.golang.org/api v0.22.0 // indirect
-	google.golang.org/genproto v0.0.0-20200611194920-44ba362f84c1 // indirect
-	google.golang.org/grpc v1.29.1 // indirect
+import (
+	"beam.apache.org/learning/katas/core_transforms/map/pardo_struct/pkg/task"
+	"context"
+	"github.com/apache/beam/sdks/go/pkg/beam"
+	"github.com/apache/beam/sdks/go/pkg/beam/log"
+	"github.com/apache/beam/sdks/go/pkg/beam/x/beamx"
+	"github.com/apache/beam/sdks/go/pkg/beam/x/debug"
 )
+
+func main() {
+	p, s := beam.NewPipelineWithRoot()
+
+	input := beam.Create(s, 1, 2, 3, 4, 5)
+
+	output := task.ApplyTransform(s, input)
+
+	debug.Print(s, output)
+
+	err := beamx.Run(context.Background(), p)
+
+	if err != nil {
+		log.Exitf(context.Background(), "Failed to execute job: %v", err)
+	}
+}
