+++ learning/katas/go/core_transforms/map/pardo_struct/pkg/task/task.go
