+++ learning/katas/go/core_transforms/map/pardo_struct/task.md
